wcsrtombs with null dst must ignore len
There's a bug in wcsrtombs (mingw-w64-crt/misc/wcrtomb.c) if null is
passed as destination buffer. In this case, wcsrtombs does not ignore
the len argument. Yet, according to the C99 standard, if the dst is
null, the len argument is ignored.

Patch by tim.kosse at filezilla-project.org