Update to autoconf v2.69 (aka automake 1.15)
This (very large) commit is solely comprised of generated files.  It comes from running "autoreconf -fiv" in all the build directories.

This commit uses the most recently released version of config.guess (which is part of autoconf).  This is a step backward from the earlier checkin which was using a beta version of config.guess.

Signed-off-by: David Wohlferd <dw@LimeGreenSocks.com>