intrin-impl.h: Fix __buildbittesti for arm
The local variables contain the full 32 bit value read from the
target, and need to be stored in 32 bit variables. The 'old' value
read is the full value of the memory location, so we need to extract
the specific bit. Mark the output registers as earlyclobber, to make
sure they don't alias the register used for input (bit).

Don't set the 'M' constraint for the bit operand, since we pass
it as a full pre-shifted integer, not the shift amount iself.

Signed-off-by: Martin Storsjö <martin@martin.st>