math: Implement the exp2() wrappers using assembly on arm64
The LLVM optimizer replaces a call to pow(2, x) back into a call to
exp2(x), which ends up in an infinite loop/recursion here. To avoid
the issue, one could either build exp2.c (or easier to accomplish, all
of libmingwex.a) using -fno-builtin or -ffreestanding, or produce the
exp2() function wrappers using assembly instead, to force a function
call to the real pow/powf functions.

Signed-off-by: Martin Storsjö <martin@martin.st>