math: Split source files with plain assembly into different directories
The files that previously had assembly for both x86 and arm are copied
into subdirectories and the arm specific parts are removed from x86 and
vice versa.

Some of the remaining .c files in the main math directory also include
arch specific assembly. But when porting to a new architecture, they
should be easier to identify (than assembly files that do assemble,
producing a symbol but no code), either via #error directives for
unsupported architectures, or via assembler errors if trying to
implicitly build inline assembly for the wrong architecture.

Signed-off-by: Martin Storsjö <martin@martin.st>