crt: Add compat wrappers for adapting the CRT startup and libmingwex to ucrtbase.dll
This intercepts the function calls that the mingw CRT startup and
other libmingwex helpers expect to use, that no longer is available on
ucrtbase.dll.

This allows using the same CRT startup object files and libmingwex both
with the normal msvcrt.dll and with ucrtbase.dll, instead of having to
rebuild them depending on which target is to be used.

Adjust the headers to make some of the inline functions static in C++
builds as well; this fixes issues with duplicate symbols where the
non-static inline function (in C++ builds) would conflict with the
CRT compat fallbacks.

Signed-off-by: Martin Storsjö <martin@martin.st>