crt: Build libwinspool.a for arm32 and arm64 as well
Use the lib64 version of the def file for this - nothing in the
def file seems like it would be architecture specific.

CMake has some linking tests where it by default links a number of
"standard libraries", and -lwinspool are among them. Make sure
this library at least exists, in order not to fail such tests
pointlessly.

Signed-off-by: Martin Storsjö <martin@martin.st>