crt: Get rid of unnecessary/unused aliases in api-ms-win-core-*.def
These (from a mingw-w64 point of view) nonsense aliases were a byproduct
of the script generating these def files. Whatever their purpose was,
it isn't how def file aliases are used within mingw def files.

GNU binutils dlltool actually doesn't seem to do anything with the part of
a line after a single '=' character, but it requires '==' to produce an
alias. llvm-dlltool on the other hand interprets both '=' and '==' the same,
producing aliases of both.

This fixes using the api-ms-win-core-* import libraries and libwindowsapp.a,
when built with llvm-dlltool.

Signed-off-by: Martin Storsjö <martin@martin.st>