headers: Default to the legacy wide string mode for ucrt
When __USE_MINGW_ANSI_STDIO is defined, we still call the custom
implementation bundled in libmingwex, which has the C99 standard
semantics for %s in wide format strings.

This matches how the wide stdio functions behave both when using
msvcrt.dll, and in all versions of MSVC, even the modern ones that
use ucrt.

Signed-off-by: Martin Storsjö <martin@martin.st>