headers: Make _{,w}find{first,next}32 the canonical name for the functions
For msvcrt.dll, the actual function is named without a -32 suffix
but the suffixed version is provided as an alias in the import
library. For msvcr*.dll and ucrtbase.dll, the actual function is
named with the -32 suffix and the unsuffixed version is an alias.
(The alias is missing in ucrtbase.dll.)

TL;DR, this makes the msvcr*.dll/ucrtbase.dll behaviour the default
and makes the old msvcrt.dll the exception (using the alias).

Signed-off-by: Martin Storsjö <martin@martin.st>