ucrtbase: Provide most common *printf functions as non-inline functions
This allows build scripts that check for functions without including
stdio.h (but just checking if the linker finds a symbol with the
given name) to detect the presence of these functions.

The asprintf and vasprintf functions get included from libmingwex,
and they are implemented in terms of _vsnprintf, so that function
must be implemented.

Signed-off-by: Martin Storsjö <martin@martin.st>