setjmp: Provide a manual implementation for arm and arm64
On arm and arm64, the setjmp/longjmp implementations (in both msvcrt.dll
and ucrtbase.dll) require binaries to have proper SEH information. On
arm, we can get around this by passing NULL as the frame/ctx parameter
to _setjmp, but that doesn't work on arm64.

Work around it by providing a custom implementation of the functions
in libmingwex instead.

The arm32 implementation is authored by André Hentschel.

Signed-off-by: Martin Storsjö <martin@martin.st>
Signed-off-by: André Hentschel <nerv@dawncrow.de>