crt: Add a new separate library libucrt.a, linking to api-ms-win-crt-*.dll instead of ucrtbase.dll
This adds def files for all the api-ms-win-crt-* dll files. I've tried to
add all the manual modifications from ucrtbase.def.in into these files, so
that linking to libucrt.a should work pretty much the same as linking to
libucrtbase.a, but ending up linking to the api sets dlls.

Signed-off-by: Martin Storsjö <martin@martin.st>