widl: Stop using AC_FUNC_MALLOC, AC_FUNC_REALLOC and AC_FUNC_STRTOD
These break cross compilation; removing them fixes cross compilation.

These not only check for the functions malloc/realloc (which are
guaranteed to be available as long as the build environment is ANSI C
compliant), but check for a "GNU libc compatible malloc".

The check for such a malloc is a runtime check, so when cross compiling,
the check just fails.

When the check for malloc has failed, all malloc calls are redirected
to rpl_malloc instead, which the project is supposed to provide as
a replacement for malloc, with the right features, but no such
fallback function exists.

Just remove these checks, and assume the ANSI C malloc/realloc/strtod
exist.

Signed-off-by: Martin Storsjö <martin@martin.st>