include/ws2tcpip.h: Do not declare non-existent `FreeAddrInfoExA()`.
In Microsoft's header `FreeAddrInfoExA` is macro defined to
`FreeAddrInfoEx`.

Note that this is purely bug-to-bug compatibility. When `UNICODE` is
defined and the ANSI variants are requested explicitly,
`FreeAddrInfoExA` will be replaced by `FreeAddrInfoEx`, and furtherly
by `FreeAddrInfoExW`. This can be observed by the following example,
which does not compile cleanly with MSVC:

    #define UNICODE 1
    #include <ws2tcpip.h>

    int main(void)
      {
        ADDRINFOEXA *ptr = 0;
        FreeAddrInfoExA(ptr);
      }

Signed-off-by: Liu Hao <lh_mouse@126.com>