math: Don't do pointer type punning via a union
Don't converting from one pointer type to another via a union;
that is still a strict aliasing violation as far as I know.

Use unions for actually reinterpreting the values instead, which
is not an aliasing violation, and is defined according to the C99
standard (which also was relied on before).

Signed-off-by: Martin Storsjö <martin@martin.st>