ucrt: Avoid using the UCRT long double nanl function on x86
On x86, mingw's 80 bit long double doesn't match MSVC's 64 bit long double.

This was done for a number of other functions in
d8ed328d4656cc467d692e1c5bd2dc7007695ef8, but the nanl function was
overlooked (in addition to all the complex functions that I've
omitted in my testing so far).

Signed-off-by: Martin Storsjö <martin@martin.st>