crt: Add compatibility wrappers for _create_locale and _free_locale in msvcrt.dll
These functions only exist in msvcrt.dll since Windows 8. They were
added (without any remarks) to the def files in b508bb87ad.

Add _CRTIMP to them in headers, to make sure all accesses go through
the __imp_ symbol, to avoid going through the wrapper function after
the firs call, which sets it to point to the real function if available,
or a stub if not.

This matches the existing wrapper for _get_current_locale, added in
9116cc881d.

Add them to the msvcrt.dll import libraries for x86_32 and x86_64,
but not for arm, as they have existed in msvcrt.dll since the first
available versions for arm. Also move the existing wrapper for
_get_current_locale from src_msvcrt (for any architecture) to the
x86 specific variables, to avoid using that unnecessary wrapper for
arm.

Signed-off-by: Martin Storsjö <martin@martin.st>