ucrt: Use multibyte functions from ucrt instead of libmingwex
Use a different definition of the mbstate_t type for ucrt, and
move related functions from libmingwex.a into libmsvcr*.a.

A few stdio functions in libmingwex use these functions; add an
override to them to, to allocate the larger form of the mbstate_t
struct, to make it work when combined with ucrt.

Signed-off-by: Martin Storsjö <martin@martin.st>