crt: Remove __chkstk from libkernel32.a for arm32
This matches what was done for x86_64 in 04f28a55c3.

On Windows 10 IoT, this symbol apparently isn't available from
kernel32, see
https://github.com/mstorsjo/llvm-mingw/issues/67#issuecomment-593431931
for discussion.

The symbol is available in the compiler-rt builtins library, like
the corresponding ones for all other architectures as well.
Normally, that one is the one included, but if -lkernel32 is
explicitly passed when linking (as is done by e.g. CMake), that one
is searched before the compiler-rt builtins library.

Signed-off-by: Martin Storsjö <martin@martin.st>