crt: Convert the arm sincos.c to assembly
This avoids optimizing the sincos function to a recursive call to
itself, if built with -ffast-math (unless built with -fno-builtin-sin).

While building with -ffast-math can break certain math routines and
thus can be considered a self-inflicted issue, it's safest to avoid
constructs that potentially can be optimized into an infinite self
recursion.

Signed-off-by: Martin Storsjö <martin@martin.st>