headers: provide the WinMain prototype to winstore apps
A proper UWP app should create a IFrameworkViewSource class, instantiate it and
call CoreApplication:Run() with it.
https://docs.microsoft.com/en-us/uwp/api/windows.applicationmodel.core

To create the instance the app still needs an entry point and that's still
WinMain:
https://stackoverflow.com/a/42211357/1266123
https://stackoverflow.com/a/53661306/1266123

So we provide the proper prototype for UWP apps as well.
This is still what the C runtime will call.

Signed-off-by: Liu Hao <lh_mouse@126.com>