headers: Hide _IO* flags in UCRT mode.
Not only does gnulib key off of the definition of _IOERR to know whether
to use its UCRT definitions, but the constants are different between
MSVCRT and the (internal) constants in UCRT source.

It turns out that the constants _IOFBF _IOLBF and _IONBF are still
present in the UCRT headers (they are used by setvbuf), but I have
confirmed none of the others are.

Reference: https://github.com/msys2/MINGW-packages/issues/8310
Signed-off-by: Jeremy Drake <jeremyd2019@users.sourceforge.net>
Signed-off-by: Martin Storsjö <martin@martin.st>