crt: Fix printf formatting of %a on arm/aarch64 with __USE_MINGW_ANSI_STDIO
The mingw_pformat.c code converts all input floats to an 80 bit
float representation and analyzes that struct when printing the
float.

This was already fixed in
a56ab0d03b5f8238527ea53f6a759f165eca0a76, but one case in
was missed. Factorize the conversion from 64 to 80 bit floats to
a function used to intialize the __pformat_fpreg_t struct.

Signed-off-by: Martin Storsjö <martin@martin.st>