crt: Move misc/seterrno.c from src_libmingwex into src_msvcrt32 and src_msvcrt64
The _set_errno function is defined in arm versions of msvcrt.dll,
and in all the newer numbered versions of msvcr*.dll, and UCRT.

This avoids issues with the recently added _vscprintf in libmsvcrt.a
which depends on _set_errno, which was present in libmingwex.a.
When linking with ld.bfd, a later library can't pull in symbols from
earlier libraries (unless specifying --start-group --end-group around
the list of those libraries).

Signed-off-by: Martin Storsjö <martin@martin.st>