headers: Don't use __mingw_wcstod/wcstof on UCRT
The UCRT provided versions should be fully C99 compliant.

This matches what was done for strtod/strtof for the initial
configuration for UCRT in e1a5c52931e5c623a461df98c0694287fcc4377c.

Also clarify the endif comment, and make the condition in wchar.h
consistent with the one in stdlib.h.

Signed-off-by: Martin Storsjö <martin@martin.st>